refactor: generator
