+++ lib/generators/subscriptable/templates/existing_subscriptable_migration.erb
