+++ lib/generators/subscriptable/templates/subscriptable_migration.erb
